+++ web/src/main/kotlin/dev/fernandocarvalho/systemdesignstudy/order/MessageSender.kt
@@ -1,4 +1,4 @@
-package systemdesignstudy.dev.fernandocarvalho.systemdesignstudy.order
+package dev.fernandocarvalho.systemdesignstudy.order
 
 interface MessageSender<V> {
     fun send(destination: String, key: String, correlationId: String, value: V)
