+++ web/src/main/kotlin/dev/fernandocarvalho/systemdesignstudy/order/business/OrderFinder.kt
@@ -1,0 +1,7 @@
+package dev.fernandocarvalho.systemdesignstudy.order.business
+
+import dev.fernandocarvalho.systemdesignstudy.order.contracts.Order
+
+interface OrderFinder {
+    fun findAll(): List<Order>
+}
