+++ web/src/main/kotlin/dev/fernandocarvalho/systemdesignstudy/order/business/OrderRepository.kt
@@ -1,0 +1,10 @@
+package dev.fernandocarvalho.systemdesignstudy.order.business
+
+import dev.fernandocarvalho.systemdesignstudy.order.contracts.Order
+
+interface OrderRepository {
+
+    fun save(order: Order): Order
+    fun findAll(): List<Order>
+
+}
