--- web/src/main/kotlin/systemdesignstudy/dev/fernandocarvalho/systemdesignstudy/order/business/OrderFinder.kt
@@ -1,7 +1,0 @@
-package systemdesignstudy.dev.fernandocarvalho.systemdesignstudy.order.business
-
-import systemdesignstudy.dev.fernandocarvalho.systemdesignstudy.order.contracts.Order
-
-interface OrderFinder {
-    fun findAll(): List<Order>
-}
